move bubble support to own file
